misc: moved functions to provider utils
